remove all the warnings in the repo connected to impulse-response solve

+++ src/geometry/BoundingBox.h
@@ -37,7 +37,7 @@ class BoundingBox {
       _minBound = FLT_MAX;
       _maxBound = -FLT_MAX;
 
-      for ( int i = 0; i < vertices.size(); i++ )
+      for ( size_t i = 0; i < vertices.size(); i++ )
       {
         if ( vertices.at( i )[0] > xmax() )
         {
